Add solution sprite
The solution sprite is essentially a compound sprite containing child
droplet sprites as well as a test tube image sprite.

We can apply tweens to the droplets to make them move based on an
activity level.

We can update the colour of the solution which tweens each rgb
component and slightly randomizes the tweening of droplets.